requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
gspread
google-auth
python-dotenv
